dedupe-tile-component * Rename DepartureTile to ChronoDepartureTile * Rename TileRows to ChronoTableRows * Import Tile and TileHeader from the new common components * Cleanup and refactor ChronoDepartureTile and ChronoTableRows. Use css-modules for styling.
