+++ src/components/DateRow/NewDayTableRow.module.scss
@@ -1,5 +1,5 @@
-.date-row {
-    border-bottom: none !important;
+.newDayTableRow {
+    border-bottom: none;
     text-transform: capitalize;
     width: max-content;
     white-space: nowrap;
